backend/chatbot/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
google-generativeai
pydantic
requests
